+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     - Removed Dataset interface
     - Neural net parameter update in Layer instead of Optimizer
     - Changed order of distance-based clusterer hyperparameters
+    - Improved cluster radius estimation in Mean Shift
 
 - 0.0.14-beta
     - Added feature importances to Gradient Boost